refactor(invoices): the payment ledger becomes its own sub-router
`server/api/invoices.ts` sat at 534 lines against a 535 cap, so the next
change to it — the partial-refund writer #19 Task 3 needs — could not add
even an import line. This is a pure move: no route path, operationId,
schema, status code or handler body changes, and the OpenAPI snapshot is
byte-identical (tests/unit/mcp all green without regeneration).

The seam is the sub-resource, not the line count. Everything under
`/{id}/payments` is one thing: append-only rows describing money that
moved, gated on the `financial` capability rather than role alone, with
its own correction mechanism (a reversing row, never an edit) and its own
rule about what reaches QuickBooks. What stays behind is the invoice ROW
— create, list, mark sent, mark paid, void, and the request-payment
orchestration. `grep -c orderPayments` now answers which file you want.

Composed with `.route('/', …)`, the shape every other split router here
uses (calendar, inspections, admin, auth). That preserves the hono/client
RPC path exactly, which is the thing a reader should doubt: the page
still calls `api.invoices[":id"].payments.$post` and `type-check:app`
proves it — the api tsconfig alone would not have.

`INVOICE_ID` moved to `invoice.schema.ts` because two route modules now
need it and importing it back from the parent would be a cycle. It also
belongs there under the repo's own rule that schemas live in
`lib/validations`, and it carries its comment with it: `invoices.id` is
opaque TEXT and a route demanding a UUID 400s a real invoice.

The baseline ENTRY is removed rather than tightened. At 339 lines the
file is governed by the ordinary 400-line rule again; re-baselining at
339 would have recreated the same wall a little further out.
